usb: typec: ucsi: improve the get current cam error message
The "Get Current Connector Alternate Mode" command is optional. The PPM
will set the "Not Supported Indicator" flag in the command status, if it
isn't supported (-EOPNOTSUPP error is reported by ucsi_exec_command()).

As this command is optional, this case shouldn't be treated as an error:
- Improve the error message to display the error code.
- Turn the not supported case into a dev_dbg instead of a dev_err.

Note: This error message is no longer seen upstream, after:
commit b9aa02ca39a4 ("usb: typec: ucsi: Add polling mechanism for partner
tasks like alt mode checking")

Signed-off-by: Fabrice Gasnier <fabrice.gasnier@foss.st.com>
Change-Id: Id7fcb5477775b90508790064d0572d310e5db5e8
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/265182
Reviewed-by: Amelie DELAUNAY <amelie.delaunay@foss.st.com>