nvmem: stm32: add warning when upper OTPs are updated
As the upper OTPs are ECC protected, they support only one 32 bits word
programming.
For a second modification of this word, these ECC become invalid and
this OTP will be no more accessible, the shadowed value is invalid.

This patch adds a warning to indicate an upper OTP update, because this
operation is dangerous as OTP is not locked by the driver after the first
update to avoid a second update.

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: Ifa806ff30d7c669ba9a3df9c6b64698a2dcc073a
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/266100
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>