remoteproc: stm32: Manage the boot address get from the elf file
On the stm32mp257 Soc, when the firmware is not delegated to the
remoteproc TEE, the non-secure vector table pointer needs to be specified.
The principle is to look for the ".isr_vectors" section in ELF file to
get the vector table address and to store it in the appropriate
HW configuration register.

Notice that the "rproc_elf_get_boot_addr" helper can not be used as
the address returned corresponds to the reset handler function (elf
entry point) not to the expected vector table address.
Moreover the strategy that consists in updating the ELF entry point to
point to the vector table is not reliable as the entry point address is
a ELF generic property(for instance used by the GDB tools to jump in
code after the firmware load).

In addition the "st,syscfg-nsvtor" property is also used to determine
if the firmware will be loaded by the Linux or the secure context.
If defined the driver considers that it has to manage the firmware.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Change-Id: Iad8ed2e33ce8761621a3397a16435e60aa491b1c
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/269211
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Eric FOURMONT <eric.fourmont-ext@st.com>
Tested-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Domain-Review: Arnaud POULIQUEN <arnaud.pouliquen@st.com>