ARM: dts: ccmp15: add support to manage FDCAN standby lines
The standby line in CAN transceiver is active at high level, so to avoid enter
in standby mode we must to manage this line. Since the linux driver does not
manage this line, this commit forces this line to low level to make sure that
the CAN transceiver is in normal mode.

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>