net: phy: realtek: add property to disable EEE for RTL8211F
Add property "realtek,eee-disable" to have possibility to disable
EEE (Energy-Efficient Ethernet) feature in Realtek Phy.
This feature is activated by default and in some usecases (TSN, gPTP etc..)
gPTP peer delays of 18us is observed when EEE is activated.
IEEE802.1AS mention to consider a system as AScapable; the peer delay
should be around 800ns. So in this situation we are hugely out of spec
and so no gPTP compliant. Without EEE, delay is improve.

Change-Id: I587aa104f8e4d45f87b18e870cdfbf222b6c9058
Signed-off-by: Christophe Roullier <christophe.roullier@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/278943
Tested-by: Eric FOURMONT <eric.fourmont-ext@st.com>
Domain-Review: Eric FOURMONT <eric.fourmont-ext@st.com>
Reviewed-by: Eric FOURMONT <eric.fourmont-ext@st.com>