ccmp15: change LDO1 (VDDA) to 3.0V and do not make it always-on
VDDA (LDO1) powers the analog block (ADC).
LDO1 is itself being powered at 3.3V, so set LDO1 voltage to 3.0V
to account for the possible max drop.

With this voltage on VDDA, we can either use the internal VREFBUF
as ADC voltage reference (up to 2.5V), or an external voltage
reference (up to 3.0V).

Remove the regulator-always-on property, since this regulator doesn't
need to be powered all the time when using VREFBUF.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

    https://onedigi.atlassian.net/browse/DEL-8216