ARM64: dts: ccimx93-dvk: use panel-lvds for LVDS interface
The standard configuration for the AUO G101EVN01.0 implemented in panel-simple
does not work well with the i.mx93 (at least with the current video
drivers). With that configuration, the colors are not properly
displayed.

To correct it, using a different data-mapping, we need to move to panel-lvds
driver which provides more configuration options.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>