ccmp1_defconfig: disable Etna Vivante driver
Originally, this driver was built as module on ST BSP but
it was later blacklisted by the meta-st-stm32mp layer.

In commit 9e5d771bf20f the Etna driver was changed to built-in
and it somewhat conflicts with the Galcore module, not
creating the node /dev/galcore which is checked by systemd
service and script 'check-gpu' resulting in a timeout message
on start and a delay of several seconds in the running of
Weston service.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>