Revert "ARM: dts: move UART7 configuration into an overlay for CCMP15"
DT overlays are not able to remove properties so this overlay
cannot really deactivate the RS485 mode.

This partially reverts commit b480eb9110a77ae09b416d7320c7a93ba33999cd.

We keep the sleep pincrtl from that commit.
We also remove the commented block for RS232 because it's the
same than RS485, except for the 'linux,rs485-enabled-at-boot-time'
property.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>