net: phy: smsc: set custom function to obtain PHY abilities on LAN87xx Ethernet PHYs
This commit fixes a race condition when the 10/100 Ethernet PHY is not
initialized in the bootloader and arrives to Linux with the ETH2_CLK clock
signal uninitialized. In the boot process, first Linux initializes the first
Ethernet MAC for the Gigabit Ethernet PHY and then initializes the second
Ethernet MAC for the 10/100 Ethernet PHY. Since CCMP13 DVK board shares the
MDIO bus #1 for both Ethernet PHYs, both PHYs are initialized at the same time
during the probe of the first Ethernet MAC, however the ETH2_CLK clock signal
is required to feed the 10/100 Ethernet PHY is not set until the second
Ethernet MAC is initialized. This causes the 10/100 Ethernet PHY probe fails
due to it is not accessible to read the PHY's abilities.

So this commit first verifies if is able to read the PHY registers, and if not,
returns an EPROBE_DEFER to try again later, then in the next try hopefully the
clock line is enabled and it will work.

https://onedigi.atlassian.net/browse/DEL-8420

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>