ARM: dts: ccmp1: add overlays for BLE Raw test
To perform the BLE HCI raw test we need the BT UART to be available
in Linux. That doesn't happen when we enable the 'cyw4373_bt' child node
in the usart.

Create a new overlay for that specific configuration, where the
bluetooth usart is enabled and the 'cyw4373_bt' child node is not. Then,
BT initialization should be performed in Linux using 'hciattach'.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>