drivers: pwm-imx-tpm: prevent applying changes if suspended
When the driver goes into suspend pm state, the clock is disabled. Under
this conditions, the pwm interface should not be accessed because it
will fail.

On the cc93 platform, when there is a display connected using the
drm_panel support, the kernel calls drm_panel_enable during the pm
resume of the system.

If the tpm driver is used as pwm-backlight, pwm_apply_state is called
before the pwm driver (pwm-imx-tpm in this case) has been resumed,
causing that the tpm registers are accessed with the clock disabled, what
produces that the device silently dies and never resumes.

This commit prevents that situation by rejecting accessing the tpm while
the drivers is suspended.

The enabled pwms (tpms) are also restored on the resume function because
they lose their values after disabling the clock.

https://onedigi.atlassian.net/browse/DEL-8494

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>