media: entity: NOT TO BE UPSTREAMED: add stream-count in media-entity struct
Add back the stream_count field of the media_entity struct in order to
count the number of time an entity has been start streamed.  DCMIPP multi-pipe
code handling needs to figure out if an entity is already streaming or not
in order to walk through the pipeline and start or stop each entities.

Change-Id: Ic6134f4b7b28be4f1b41b1b199bf60ae6a206a53
Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>