arm64: dts: st: add sdmmc2 & 3 nodes on stm32mp257f-ev-revB [MP25_REVA]
The sdmmc2 node is used for the eMMC embedded on the board.
The sdmmc3 node is used to connect an SD-card with an extension board
connected to RPi GPIO expansion connector. As the card is not present
by default, the node status is disabled.

Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Change-Id: I9656e577c5823c9fa4de2c3e4cb4439eb11a1720
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/308419
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>