arm64: dts: st: Add PCIe EP on the stm32mp257f-ev-revB board
Configure PCIe EP and the ComboPHY with external pad clock on
eval-revB board to be used as a device. Exclusive with pcie_rc
that is the default.
The EP is configured for CommonRef clock, need to deassert CLKREF#.
Note that the HW board needs to be modified to root the PCIE_CLK pins
to PCIE_CLKIN.

Signed-off-by: Christian Bruel <christian.bruel@foss.st.com>
Change-Id: I6dcffdd22eee2207ed0e78f7c9ce6dcfb4ca5407
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/321968
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Fabrice GASNIER <fabrice.gasnier@foss.st.com>
Domain-Review: Fabrice GASNIER <fabrice.gasnier@foss.st.com>