brcmfmac: kso sequence sleep/delay enhancement
In the current implementation of kso sequence in fmac, there is a sleep
in the very first write and read check. So even if the device is
already up, it still sleeps for that particular duration which leads to
more power consumption for both Host and Device as well as increased
response delays.
Changes:
1) Remove the usleep_range between the first write and read check
2) Introduce a delay after first write-read check, the value of delay
   will be considered based on try_cnt
3) Replace udelay with usleep_range as its a recommended method instead
   of busy-wait of udelay

Signed-off-by: Ishan Deshpande <IshanShirish.Deshpande@infineon.com>