ov5640: fix initialization on ccimx93
The driver in v6.1.22 (upstream + NXP fixes) changed heavily since
5.15. Our old workaround to adjust the PLL multiplier has now some bad
effects (like wrong saturated colors), so remove that completely.

The powerup sequence also changed. The driver now handles two gpios
(powerdown and reset), and the reset gpio is not toggled unless you have
the powerdown gpio defined. So, as a workaround, change the ccimx93 device
tree to define the powerdown gpio which seems to be enough at the moment
to get the camera detected.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>