arm64: dts: stm32: add rpmsg i2c stm32g0-typec device for stm32mp257f-ev1-revB boards
Add the support of the stm32go typeC device on the rpmsg I2C adapter.
Instead of communicating with an external G0 processor, it communicates
with the coprocessor relying on a virtual i2c. G0  phuysical interrupts
are replaced by a rpmsg virtual interrupts controller through the
irq-rpmsg service.
notice that this impose that the rpmsg framework treat message under
interrupt context. this solution should be replaced by virtio-gpio and
virtio-i2c solutions.

Signed-off-by: Pankaj Dev <pankaj.dev@st.com>
Change-Id: I054154c603ad1567470b0e760cc8f76b1a6411a9