dt-bindings: phy: stm32-femtophy: add property to select Internal VBUS comparator
Boolean property to select Internal VBUS comparator present
inside the femto-phy. To be used for the scenario when the VBUS
presence is to be detected by the PHY, using its internal comparator,
and is notified to usb-ctrl via UTMI signal. The opposite scenario
is when VBUS presence is detected by some external comparator and
usb-ctrl is notified using VBUSVLDEXT

Signed-off-by: Pankaj Dev <pankaj.dev@st.com>
Change-Id: I33f88627a653c16dc63798679e07372680193738