hsio: pcie: set fixed clock for hsio interface
The "gpio-gate-clock" requires an enable-gpios property,
without that property, the clock-hsio doesn't work
and raise this error:

clock-hsio-refa: gpio_clk_driver_probe: Can't get 'enable' named GPIO property

In the kernel 5.4 despite of this error, everything was
working fine but in kernel 5.15 the PCI interface is not loaded.

This gpio in the hsio_refa/refb node is an output that request
a clock. Taking a look to the reference design of NXP, this GPIO
is connected to an input to the 100 MHz crystal to enable that signal,
in our case that signal comes from a always enable 100 MHz crystal,
so we don't need to have a gpio-gate-clock, instead of doing that
a fixed clock is a better solution.

Following our schematic, this gpio4 1 is configured as digital output
and it is connected to the qca6574 wireless chip to an output that
is doing the same function, request a clock that is always provided
in our case, so we are connecting two digital outputs with a pull-up resistor.
In this case these GPIOs once the wireless driver is loaded both
went to down, and we are not over driving the GPIO driver (also we have
a pull up in the line as protection).

So that GPIO is not needed by the driver wireless and we can free it
from the clkreq-gpio property, and to avoid the connection
of the two outputs, the GPIO4 1 is now configured as input with
a pull-up when the wireless overlay is used.

This error comes from: 5c5de3e6c1f0500014f3966295ebcd39357e733d

https://onedigi.atlassian.net/browse/DEL-8605

Signed-off-by: Francisco Gil <francisco.gilmartinez@digi.com>