ARM64: dts: ccimx8x:  add missing properties to reserved-memory node
In Linux v5.4 and older, the initial definition of the reserved-memory node
used to be in the imx8x SoC .dtsi, but it has since been moved to the imx8x
MEK's .dtsi. Because of this, our node was missing essential properties which
caused the regions to not be parsed properly.

Add the "address-cells", "size-cells" and "ranges" properties to our node to
fix this.

https://onedigi.atlassian.net/browse/DEL-8604

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>