ARM64: dts: ccimx8x: match memory regions with the ones in the MEK dts
In order to avoid unpredicted behavior, make our memory regions match the ones
in the MEK's device tree. The new CMA address used in the MEK (0xC0000000)
assumes a RAM of over 2 GiB, which isn't the case in our modules, so keep
using the old address.

Note that NXP now uses a separate memory region for the GPU, but said region is
only available in targets with more than 2 GiB of RAM. Keep using our mechanism
for now.

https://onedigi.atlassian.net/browse/DEL-8604

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>