pinctrl: stm32: add data flag need_syscon
The pinctrl driver writes directly in the device specified by the
property 'st,syscfg' to program the interrupt mux.
The mux is inside the EXTI for MPUs and inside the st,stm32-syscfg
for the MCUs.

To prepare for EXTI to handle the mux directly, tag the pinctrl
for MCUs with the flag need_syscon; it indicated that the device
needs the property 'st,syscfg' to access st,stm32-syscfg. The MPU
devices don't need the property, but tolerate it for backward
compatibility with old DTs.

Signed-off-by: Antonio Borneo <antonio.borneo@foss.st.com>
Change-Id: I5120c88e7504444bb774e10b7d8c3c59cda092aa
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/329523
Domain-Review: Amelie DELAUNAY <amelie.delaunay@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Alexandre TORGUE <alexandre.torgue@foss.st.com>