brcmfmac: Send CSA to supplicant on channel change
Fixes: SWLINUX-3649

When the AP changes channels, the underlying Driver/FW will negotiate
correctly and stay connected, however, the Driver never informs netlink
of the channel change. When the channel is queried, the channel is
correct, however, wpa_supplicant is waiting on the
NL80211_CMD_CH_SWITCH_NOTIFY event, which is triggered by
cfg80211_ch_switch_notify which should be called by the FMAC driver.

This patch adds support to send CSA event to wpa_supplicant.

Signed-off-by: Pradeep Gurumath <pradeep.gurumath@infineon.com>