kbuild: keep clean the folder for external device tree
The current build system builds the dtbs inside the folder of the
external DT then, depending on ARCH, copies the dtbs either in
arch/$(SRCARCH)/boot/dts/st on in arch/$(SRCARCH)/boot/dts or in
the corresponding folders in case of out-of-tree build.

Change the build system to always build the dtbs and the associated
temporarily files directly in the final folder.
This prevents the copy of the dtbs and keeps clean the folder of
the external DT, either in in-tree and in out-of-tree builds.

Signed-off-by: Antonio Borneo <antonio.borneo@foss.st.com>
Change-Id: Idd71a14b5199bdf645d9b2814a498db433fcd86c
Depends-On: If58e077fedea6557e0a69ff04966e40af13b8321
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/329554
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Amelie DELAUNAY <amelie.delaunay@foss.st.com>
Tested-by: Amelie DELAUNAY <amelie.delaunay@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Domain-Review: Amelie DELAUNAY <amelie.delaunay@foss.st.com>