ccimx8m: remove processing of 'rs485-rts-active-high' DT property
The serial core supports processing RS485 DT properties. One of these
properties is a boolean one that establishes the polarity of the DE signal.
It is called 'rs485-rts-active-low'. The presence of it reverses the
default polarity, which is active high.
In other words, for an active high polarity, we don't need a new property.
This commit removes the exisiting property from ccimx8mnano and ccimx8mmini
device trees and the custom code that parsed it in the driver.
Besides, the RTS line is used as GPIO in these platforms and the polarity
is decided in rts-gpios property, anyway.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>