brcmfmac: avoid memory use-after-free in the escan timeout worker function
With a WLAN chip plugged in over USB Bus, if an ongoing scan operation is
interrupted with a sudden USB unplug, the brcmf_cfg80211_info struct memory
would be freed, the pointer is set to NULL and the escan_timeout timer is
cancelled if not fired yet. There is a rare chance for escan_timeout timer
to have already fired and scheduled the escan_timeout_work, just before the
USB unplug action frees the struct memory. When the scheduled work executes
the already freed memory would get accessed.

During this race condition, avoid this memory user-after-free error at the
escan timeout worker function, by doing a NULL check and return.

This Fixes CVE-2023-47233

Signed-off-by: Gokul Sivakumar <gokulkumar.sivakumar@infineon.com>