rtc-mca: fix mca fw version where prepare command is supported
Commit f764cd6f8295 was enabling prepare command only for fw versions that support
it; but was setting that version to MCAv1.20 when in fact rtc prepare command was
implemented in MCAv1.19.

Prepare command is required when i2c bus speed is higher than 100KHz.
Other commits limit i2c bus speed to 100KHz for MCA versions lower than 1.19,
but let 400KHz to MCAv1.19 and above.
This is making MCAv1.19 fail reading date and alarm as it works at 400KHz but
without prepare command.

https://onedigi.atlassian.net/browse/DEL-8806

Signed-off-by: Hector Bujanda <hector.bujanda@digi.com>