ARM: dts: ccmp1: mfg: reduce Wireless SDIO clock to 1MHz
Wireless SDIO lines are available outside of the SOM. On the Functional
test board, these lines are tested as GPIO loopbacks in the non-wireless
modules. This connection has an effect on the SDIO interface of wireless
modules, where the clock line is very sensitive.

Reduce the Wireless SDIO clock to prevent the wireless initialization
from failing.

https://onedigi.atlassian.net/browse/DEL-8775

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>