ccimx93: pca9450: disable WDOG_B only on suspend state
commit 0fd221cbc2eef6bd0414bdd57cdfdd947b4870e0 added support to ignore
the WDOG_B pin unconditionally.
That pin is used by the CPU to notify the PMIC to be reset and therefore
reboot the system, so we must only disable it before going to suspend mode.
Therefore, disable the WDOG_B pin signal while entering in suspend
state and restore the value when resuming.
While on it, change the variable name to be more accurate, and update
the binding and affected device trees.

https://onedigi.atlassian.net/browse/DEL-8328

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>