brcmfmac: fixing mbss iovar setting for multiple BSS is enabled
1.) if the user enabled hostapd on wlan0,
the type of wlan0 changes to AP_TYPE.

2.) then the user uses iw command to created a new interface wlan1,
or wlan2 to AP mode.
the vif->mbss belongs to wlan1, and wlan2 will always is "1"

which causes while the start_ap is called,
it highly relies on the interface's index to set related AP's parameter.

if the first enabling AP interface is not the "first"
interface with vif->mbss is "0",
the upcoming ap interface enabling will fail.

for example,
- hostapd_cli -p /var/run/hostapd -i wlan2 disable
- hostapd_cli -p /var/run/hostapd -i wlan1 disable
- hostapd_cli -p /var/run/hostapd -i wlan0 disable

- hostapd_cli -p /var/run/hostapd -i wlan2 enable  //pass
- hostapd_cli -p /var/run/hostapd -i wlan1 enable  //fail
- hostapd_cli -p /var/run/hostapd -i wlan0 enable  //pass

to make the judgement of check_vif_up can be workable.

the command of disable shall add "ifconfig xxx down".
for example,
- hostapd_cli -p /var/run/hostapd -i wlan2 disable
- ifconfig wlan2 down
- hostapd_cli -p /var/run/hostapd -i wlan1 disable
- ifconfig wlan1 down

Signed-off-by: Carter Chen <carter.chen@infineon.com>