ARM64: ccimx93-dvk: add sleep pinctrl for Ethernet interfaces
On suspend, there are occasions in which a spurious voltage can be sensed
at the 1V8_PHY_N lines even when the PHY power (coming from BUCK4) is
switched off.
Changing the IOMUX of the Ethernet lines to GPIO and selecting an
appropriate padding control avoids this occasional voltage feedback.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>