Revert "LF-9151 soc: imx: Increase the auto clock gating idle strap on imx93"
Having higher values of the auto clock gating idle period is known to cause
flickering on some HDMI displays. Revert the default value back to how it
was originally to avoid flickering and improve power consumption. Note that
this parameter can be configured at runtime, so it can be changed on the fly
if needed for use cases that require better DDR latency.

This reverts commit 94cd8244d97a050027e13d1848283c93344a1405.

https://onedigi.atlassian.net/browse/DEL-8843

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>