brcmfmac: set authorized flag in CONNECT event for all 4way HS offload case
The 2.11-devel wpa_supplicant expects the driver to notify userpsace about
port authorization after successful 4-way handshake completion if it was
offloaded to the Driver.

Set the authorized flag in CONNECT event even for Open/SAE Authentication
case along with the 802.1X authentication case. The cfg80211 on receiving
connect result event with this flag, adds the NL80211_ATTR_PORT_AUTHORIZED
attribute in the NL80211_CMD_CONNECT event sent to userpsace.

Signed-off-by: Gokul Sivakumar <gokulkumar.sivakumar@infineon.com>