non-upstream: Increase delay in kso sequence
There exists a corner case specially when in low RSSI, where at DS2
exit, SDIO core accesses from ucode and FMAC clash and DS2 exit fails
leaving the chip in a bad/stuck state.

Changes: Increase the usleep_range to (2500 to 3000)us when the first
KSO wr-rd-check fails before starting subsequent checks to avoid the
clash. This gives some more time for device to wakeup.

NOTE: This is a WAR and will be updated once proper fix is found.

Signed-off-by: Pranav Vinayak Wagh <PranavVinayak.Wagh@infineon.com>