brcmfmac: Revert "wireless: brcmfmac: Use netif_rx()."
This reverts commit a791491c62e4 ("wireless: brcmfmac: Use netif_rx()"), due to
the changes introduced to add a local_bh_disable() section in netif_rx() to
ensure softirqs are handled if needed in commit baebdf48c3600 ("net: dev: Makes
sure netif_rx() can be invoked in any context."), started to be part of the
kernel in v5.18. So the netif_rx in the older kernel version than v5.18
(such as 5.15) doesn't use local_bh_disable to make sure the netif_rx_internal
is protected and make sure the incoming packet is processed completely.

https://onedigi.atlassian.net/browse/DEL-8667