arm64: dts: ccimx8x: update mfg board device tree
Define the MIPI Analyzer as a simple panel under the 'mipi0_dsi_host' and
'mipi_dsi1_host' nodes.
Timing configurations are defined in the Kernel driver.

Configure High drive mode in every pin of the Parallel Camera interface
to have a more stable communication.

Also, rearrange GPIO definition to match the pairs as represented in the
schematic.

https://onedigi.atlassian.net/browse/DEL-8934

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>