gpmi-nand: use original NXP fork implementation of common_nfc_set_geometry()
Originally, up until NXP's lf-5.15.71-2.2.2 BSP release, the logic needed to
change the BCH layout setting depending on the oob size wasn't available in
mainline Linux, it was added in commit 3bfb219e311ba0780a4a8e88af39e07d67836132
as part of NXP's fork. This was also true for U-Boot.

Starting in Linux v5.19 (d9edc4bc67c4fd77b30d0a55bc84b775c2f00bab) and U-Boot
v2020.07, the logic was upstreamed, although changed slightly to use legacy
geometry by default. This change in logic causes the following error to appear
repeatedly when booting a ccimx6ul:

    ubi0 warning: ubi_io_read: error -74 (ECC error) while reading 64 bytes from PEB 2:0, read only 64 bytes, retry
    ubi0 warning: ubi_io_read: error -74 (ECC error) while reading 64 bytes from PEB 2:0, read only 64 bytes, retry
    ubi0 warning: ubi_io_read: error -74 (ECC error) while reading 64 bytes from PEB 2:0, read only 64 bytes, retry
    ubi0 error: ubi_io_read: error -74 (ECC error) while reading 64 bytes from PEB 2:0, read 64 bytes
    CPU: 0 PID: 1 Comm: swapper Not tainted 6.1.55-dey-dey+gff796429549f+p0 #1
    Hardware name: Freescale i.MX6 Ultralite (Device Tree)
     unwind_backtrace from show_stack+0x10/0x14
     show_stack from dump_stack_lvl+0x24/0x2c
     dump_stack_lvl from ubi_io_read+0x158/0x32c
     ubi_io_read from ubi_io_read_ec_hdr+0x44/0x1e8
     ubi_io_read_ec_hdr from scan_peb+0x6c/0x788
     scan_peb from ubi_attach+0x23c/0x384
     ubi_attach from ubi_attach_mtd_dev+0x568/0xbf4
     ubi_attach_mtd_dev from ubi_init+0x188/0x238
     ubi_init from do_one_initcall+0x48/0x230
     do_one_initcall from kernel_init_freeable+0x190/0x1f4
     kernel_init_freeable from kernel_init+0x18/0x12c
     kernel_init from ret_from_fork+0x14/0x28
    Exception stack(0x90815fb0 to 0x90815ff8)
    5fa0:                                     00000000 00000000 00000000 00000000
    5fc0: 00000000 00000000 00000000 00000000 00000000 00000000 00000000 00000000
    5fe0: 00000000 00000000 00000000 00000000 00000013 00000000

We can fix this error with minimal changes by reverting the implementation of
common_nfc_set_geometry() back to how it was in the lf-5.15.71-2.2.2 fork. We
have yet to test if using a newer U-Boot with the updated BCH/OOB logic along
with the v6.1 kernel fixes the issue, but for now, apply this fix.

https://onedigi.atlassian.net/browse/DEL-8967

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>