ARM64: dts: ccimx93-dvk: add comments on disabling DMA for UARTs
iMX93 A0 revision has defects on the DMA that affect the
behavior of the UARTs. If using this revision of the CPU
the DMA must be disabled for the UARTs.
Unfortunately the removal of properties cannot be done on
an overlay. This commit adds the lines commented out, so
that they can easily be uncommented (and DMA disabled) by
users of A0 revision.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DEL-8888